models/requirements.txt
aenum==3.1.15
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.1
filelock==3.18.0
fiona==1.10.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jsonargparse==4.38.0
kagglehub==0.3.11
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lightly==1.5.19
lightly-utils==0.0.2
lightning==2.5.1
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
munch==4.0.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pretrainedmodels==0.7.4
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pycocotools==2.0.8
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.3
rich==13.9.4
rtree==1.4.0
safetensors==0.5.3
scipy==1.15.2
segmentation_models_pytorch==0.4.0
sentry-sdk==2.26.0
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboardX==2.6.2.2
timm==1.0.15
torch==2.6.0
torchgeo==0.7.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typeshed_client==2.7.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.9
wheel==0.45.1
yarl==1.19.0
